+++ _typos.toml
@@ -8,4 +8,4 @@ ba = "ba" # 7zコマンドの-baオプション
 commitish = "commitish" # softprops/action-gh-releaseのオプションの１つ
 
 [files]
-extend-exclude = ["package-lock.json", "src/store/project.ts", "*.svg"]
+extend-exclude = ["package-lock.json", "src/domain/project/index.ts", "*.svg"]
